Add VBR functions
The script now displays the extracted VBR information, but it is not complete. The cluster 2 calculation needs to be changed to handle FAT16 and FAT32 differently and the script needs to calculate the number of partitions as opposed to assuming that there will always be four partitions in an image.